+++ packages/plugins/LoadingIndicator/src/components/LoadingIndicator.vue
@@ -39,6 +39,9 @@ export default Vue.extend({
   computed: {
     ...mapGetters('plugin/loadingIndicator', ['showLoader', 'loaderStyle']),
   },
+  mounted() {
+    this.$store.dispatch('plugin/loadingIndicator/setupModule')
+  },
 })
 </script>