+++ packages/plugins/LoadingIndicator/src/store/index.ts
@@ -6,6 +6,7 @@ import { PolarModule } from '@polar/lib-custom-types'
 import { LoadingIndicatorGetters, LoadingIndicatorState } from '../types'
 import getters from './getters'
 import { getInitialState } from './state'
+import mutations from './mutations'
 
 export const makeStoreModule = () => {
   const storeModule: PolarModule<
@@ -14,17 +15,13 @@ export const makeStoreModule = () => {
   > = {
     namespaced: true,
     state: getInitialState(),
-    mutations: {
-      addLoadingKey(state, key: string) {
-        state.loadKeys = new Set([...state.loadKeys, key])
-      },
-      removeLoadingKey(state, key: string) {
-        const loadKeys = new Set(state.loadKeys)
-        loadKeys.delete(key)
-        state.loadKeys = loadKeys
-      },
-    },
+    mutations,
     getters,
+    actions: {
+      setupModule({ rootGetters, getters, commit }) {
+        commit('setLoaderStyle', (rootGetters.configuration.loadingIndicator || {}).loaderStyle)
+      }
+    },
   }
 
   return storeModule