+++ packages/plugins/LoadingIndicator/src/store/state.ts
@@ -2,4 +2,5 @@ import { LoadingIndicatorState } from '../types'
 
 export const getInitialState = (): LoadingIndicatorState => ({
   loadKeys: new Set(),
+  loaderStyle: 'v-progress-linear'
 })