+++ packages/plugins/LoadingIndicator/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## unpublished
 
 - Refactor: Remove redundant prop-forwarding by only using one component.
+- Feature: add new options parameter `loaderStyle` to choose between different loader styles.
+- Feature: add new mutation `setLoaderStyle` to choose between different loader styles at runtime.
 
 ## 1.1.0